backend/requirements.txt
fastapi==0.115.8
strawberry-graphql[fastapi]==0.259.0
uvicorn[standard]==0.34.0
pydantic==2.10.6
